Read me
## Yelp Recommendation system
Yelp contains reviews about a particular restaurant, and also it helps user to choose the restaurants nearby based on location. But it doesn't recommend any restaurants to users. In this project, text reviews are used to recommend restaurants to the users. Based on rating predictions recommendation models are developed. Users restaurants preferences can be decoded based on the users reviews about restaurants that they have visited previously.
## Description of the Dataset
The dataset is obtained from yelp.com, which consists of 1.6M reviews by 366K users for 61K businesses. The dataset consists of 5 files - bussiness, review,user,check -in and tip. Bussiness and review files are mainly used for this project. All the files are in JSON format.